Update to use raspvid

+++ install.sh
@@ -1,9 +1,12 @@
-# http://www.linux-projects.org/modules/sections/index.php?op=viewarticle&artid=14
-wget http://www.linux-projects.org/listing/uv4l_repo/lrkey.asc && sudo apt-key add ./lrkey.asc
-echo "deb http://www.linux-projects.org/listing/uv4l_repo/raspbian/ wheezy main" >> /etc/apt/sources.list
 sudo apt-get update
-sudo apt-get install uv4l uv4l-raspicam uv4l-raspicam-extras libav-tools
+sudo apt-get install libav-tools python
 
-echo "Please enter your justin-tv key"
-read key
-echo "$key" >~/.twitch_key
+if [ ! -e "~/.twitch_key" ]; then
+  echo "Please enter your justin-tv key"
+  read key
+  echo "$key" > ~/.twitch_key
+fi
+
+wget http://downloads.sourceforge.net/project/s3tools/s3cmd/1.5.0-beta1/s3cmd-1.5.0-beta1.tar.gz
+tar xvfz s3cmd-1.5.0-beta1.tar.gz
+
